Controlls all four motors at once
Open up the Serial Monitor (Ctrl + Shift + M) and you'll see the first line typed out and then the option to enter one of two commands (it's a work in progress). X will stop whatever is running (like ctrl+c in the command prompt in windows), and N allows you to reset the serial monitor and be allowed to enter a new time. After the program is done writing out the introductory text, it will give you a prompt to enter a time. Enter the time you want the motors to be on in milliseconds, and it will run for this time.